+++ changelog.md
@@ -8,6 +8,9 @@
 - Fixed command ``/heal <player>`` and `/feed <player>` if you use it in the console [#90](https://github.com/Maxlego08/zEssentials/issues/90)
 - Added two types of home usage MAX and STACK [#84](https://github.com/Maxlego08/zEssentials/issues/84)
 - Fixed home delete inventory [#89](https://github.com/Maxlego08/zEssentials/issues/89)
+- Fixed item display when you try to display an empty item
+- Fixed tabulation with no argument needed
+- Fixed message when you create a home
 
 # 1.0.1.0